view: do not draw unmapped views from offscreen buffer
So far, core has kept a copy of the unmapped view's contents in an
offscreen buffer, so that it can be drawn when the view is unmapped.
With scenegraph, the animate plugin can make and draw this copy itself,
so there is no need to put this plugin-specific functionality in core.