Fix "No free output buffer slot" messages
The problem here was that moving the cursor during repaints was sometimes
causing all free buffer slots in wlroots to be used because we were calling
wlr_cursor_set_surface() more than once per frame. Call this on idle so that
we only do it once per frame and avoid using all the free buffer slots.