wlr-surface-node: avoid sending excessive redraws
When we have a pending transaction, we want to make sure we send a frame
done to the client, so that it can resize. However, if it is already
visible on an output, we should not send additional frame done events.
Instead, we should wait until the 'normal' frame done events are sent as
part of the rendering loop.