output-layout: Don't initialize output position with values unconditionally
When initializing output::position, passing x and y arguments means that
the position will not be marked automatic, even if it's set to auto or
the default value. This spells trouble later, when enabling an output
that was previously disabled that is not the default 0,0 output. This
commit checks the position is set to automatic before blindly supplying
x and y values to the position.

Fixes #2147.