view-impl: separate map state from priv->wsurface being set
In Xwayland, it is possible to get situations where priv->wsurface is
dangling because we delay unmap until next idle.

So what happens is that a surface is unmappad and then dissociated,
while the unmap transaction is still pending. In such cases it
appears as if priv->wsurface is set but it can be dangling.