output-layout: make NOOP-1 output larger.
When outputs are unplugged, views are temporarily moved to the
NOOP-1 output. Because this output is currently only 1280x720,
windoes moved to it get moved and cropped to match its small
screen size.

Make the output 4k so windows stay where they are unless the user
is running a display larger than 4k.

This does not affect the WL-x outputs used when running in
headless mode, which continue to be created as 1280x720 by
wlroots code (attempt_headless_backend).