move: Check if grid is enabled before activating edges on drag
Before, dragging to an edge with move enabled and grid disabled would
result in the indicator being drawn for the slot but always maximized.
Fix this by emitting a signal that grid or another plugin can tell move
the request is handled, so move can try gridding. In this implementation,
we just hard code top to maximize if grid is disabled.

Fixes #2403.