Link all symbols from wf-json library
When using i.e. wf::ipc_activator_t in external plugins, they will
not load due to missing symbols such as this:

libpixdecor.so: undefined symbol: _ZNK2wf16json_reference_t9is_uint64Ev

This is due to the fact that wayfire/plugins/ipc/ipc-activator.hpp
ultimately includes subprojects/wf-json/wayfire/nonstd/json.hpp
which houses declaration of functions but not the definitions.
The definitions are in the static library wf-json. However, if
we do not link whole, even though external plugins still compile,
they fail to load at runtime. Linking whole means the symbols are
avaialable at runtime because they are included in the wayfire
binary.

Note that we cannot merely use json.link_as_whole() in wayfire
dependencies list, as we will get multiple definition erros at
link time. So, we directly include it in the dependencies for
the wayfire static library, and this fixes the problem.